backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
websockets>=12.0.1
httpx==0.27.0
python-dotenv==1.0.1
orjson>=3.9.0
scikit-learn>=1.3.0
xgboost>=2.0.0
